refactor(customBotton): 커스텀 버튼을 생성하는 위치 변경
